requirements.txt
# AI Job Hunter - Dependencies
# Python 3.10+ recommended

# Scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# PDF extraction
pdfplumber>=0.10.0

# OpenAI
openai>=1.12.0

# Geography / distance
geopy>=2.4.0

# Config
python-dotenv>=1.0.0

# Optional: LinkedIn (uncomment if needed)
# playwright>=1.40.0
